Move light sensor scheduler to its own file
