+++ components/functions/functions/chicken_door/ChickenDoor.hpp
@@ -10,8 +10,8 @@
 #include <functions/ScheduledTransitionLoop.hpp>
 
 #include <peripherals/api/IDoor.hpp>
-#include <peripherals/api/ILightSensor.hpp>
 #include <utils/scheduling/CompositeScheduler.hpp>
+#include <utils/scheduling/LightSensorScheduler.hpp>
 #include <utils/scheduling/OverrideScheduler.hpp>
 
 using namespace farmhub::peripherals::api;
@@ -42,57 +42,6 @@ public:
      * @brief Until when the override state is valid.
      */
     Property<time_point<system_clock>> overrideUntil { this, "overrideUntil" };
-};
-
-struct LightSensorSchedule {
-    Lux openLevel;
-    Lux closeLevel;
-};
-
-struct LightSensorScheduler : IScheduler {
-    LightSensorScheduler(const std::shared_ptr<ILightSensor>& lightSensor)
-        : lightSensor(lightSensor) {
-    }
-
-    void setTarget(std::optional<LightSensorSchedule> target) {
-        this->target = target;
-    }
-
-    const char* getName() const override {
-        return "light";
-    }
-
-    ScheduleResult tick() override {
-        if (!target) {
-            return {
-                .targetState = {},
-                .nextDeadline = {},
-                .shouldPublishTelemetry = false,
-            };
-        }
-        const auto& target = *this->target;
-
-        auto targetState = calculateTargetState(lightSensor->getLightLevel(), target);
-        return {
-            .targetState = targetState,
-            .nextDeadline = 1min,
-            .shouldPublishTelemetry = false,
-        };
-    }
-
-private:
-    static std::optional<TargetState> calculateTargetState(Lux lightLevel, const LightSensorSchedule& target) {
-        if (lightLevel >= target.openLevel) {
-            return TargetState::Open;
-        }
-        if (lightLevel <= target.closeLevel) {
-            return TargetState::Closed;
-        }
-        return {};
-    }
-
-    std::shared_ptr<ILightSensor> lightSensor;
-    std::optional<LightSensorSchedule> target;
 };
 
 class ChickenDoor final
